Separate GOST engine and GOST library, make GOST engine a module
The GOST engine is a dynamically loadable module rather than a shared
library, so make that explicit.  However, the programs gost12sum,
gostsum and gostsum12 need to link against the algorithms implemented
as part of the engine, so separate those out into a static library
that both programs and engine link with.