Special branch reuires patching OpenSSL
This branch is created for experiments with the patched OpenSSL version.
The patch implements Russian GOST TLS 1.2 and TLS 1.3 support in
OpenSSL.

Some parts of the patch are already included in OpenSSL 3.0, some are
not and possibly, will never be.

This branch is recommended for ditribution builders and those who want
make experiments. It MUST NOT be used instead of system OpenSSL.