test: parallel testsuite runs failed due to reused filenames
Cmake can run tests in parallel and does so for the perl tests. These
are actually the same tests twice but in "engine" and "provider"
variants.

These tests contain something like the following:
- create testdata.dat
- write to it
- use the file
- unlink the file

When two of these run in parallel, a likely scenario is the file is
unlinked by one while still being needed by the second.

This commit simply prefixes the filenames with "$ARGV[0]-" which is
either "engine-" or "provider-". With unique names, there are no more
overlaps.

PS: I had first looked for ways to disable parallel test execution (when
I thought it was perl running tests in parallel rather than cmake) but
found no way to do that in CMakeLists.txt, only through an environment
variable or by changing the command-line invocation, both of which where
far less appropriate both for distribution patches and upstreaming.