update to latest version of libprov
Recently, we saw a failure in openssl tests when building gost-engine:
https://github.com/openssl/openssl/actions/runs/14192630435/job/39768534802

It occurs because a gost-engine dependency (libprov) requires a minimum
cmake version of 3.0.  Recent distro updates have moved to cmake 4.0,
which error out if a CMakeLists.txt file requires a version less than
3.5.

The libprov tree has been updated to require 3.18 now (matching the
requirement for gost-engine), so we need to update the libprov submodule
here to pull that update in.